improve remote_peer_t

+++ include/remote_peer.hpp
@@ -6,11 +6,11 @@
 #ifndef REMOTE_PEER_HPP
 # define REMOTE_PEER_HPP 1
 # include <inttypes.h>
-# include <stdint.h>
+# include <stddef.h>
 # include <time.h>
 struct remote_peer_t {
-  time_t   seen;
-  intmax_t cent; // config entry
+  time_t seen;
+  size_t cent; // config entry
 
   remote_peer_t() noexcept;
   remote_peer_t(const size_t cfgent) noexcept;
